+++ debug.h
@@ -4,6 +4,6 @@
 #include "chunk.h"
 
 void disassemble_chunk(const struct chunk *chunk, const char *name);
-int disassemble_instruction(const struct chunk *chunk, int offset);
+size_t disassemble_instruction(const struct chunk *chunk, size_t offset);
 
 #endif // CLOX__DEBUG_H_
