+++ object.h
@@ -22,13 +22,13 @@ struct obj {
 
 struct obj_string {
         struct obj obj;
-        int length;
+        size_t length;
         char *chars;
         uint32_t hash;
 };
 
-struct obj_string *take_string(char *chars, int length);
-struct obj_string *copy_string(const char *chars, int length);
+struct obj_string *take_string(char *chars, size_t length);
+struct obj_string *copy_string(const char *chars, size_t length);
 void object_print(struct value value);
 
 static inline bool is_obj_type(struct value v, enum obj_type type)
